perf(charts): revise the d3 charts library
- [x] break down the chart examples component into presentational components;
- [x] fix the force directed chart - fix the links count value mapping;
