+++ libs/client-d3-charts/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@rfprodz/client-d3-charts",
-  "version": "1.3.16",
+  "version": "1.3.17",
   "description": "Angular chart components based on D3JS (https://d3js.org).",
   "keywords": [
     "angular-charts",
