+++ libs/client-d3-charts/src/lib/components/chart-examples-pie/chart-examples-pie.component.html
@@ -1,0 +1,3 @@
+<div class="container" *ngIf="pieChartConfig$ | async as config">
+  <app-pie-chart [chartId]="'pie-example-1'" [data]="config.data" [options]="config.options"></app-pie-chart>
+</div>
