+++ libs/client-d3-charts/src/lib/components/chart-examples-radar/chart-examples-radar.component.scss
@@ -1,0 +1,13 @@
+:host {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+
+  .container {
+    display: flex;
+    flex: 1 1 auto;
+    flex-wrap: wrap;
+    width: 100%;
+  }
+}
